requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.23
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
alembic>=1.13.1
requests>=2.31.0
email-validator>=2.0.0

# Chat System Dependencies
openai>=1.0.0
cryptography>=41.0.0


 